+++ lib/src/base/platform_streams.dart
@@ -17,12 +17,16 @@ class PlatformStreams {
 
   bool isInitialised = false;
 
-  void init() {
+  Future<void> init() async {
     _currentDurationController =
         StreamController<PlayerIdentifier<int>>.broadcast();
     _playerStateController =
         StreamController<PlayerIdentifier<PlayerState>>.broadcast();
-    AudioWaveformsInterface.instance.setMethodCallHandler();
+    _extractedWaveformDataController =
+        StreamController<PlayerIdentifier<List<double>>>.broadcast();
+    _extractionProgressController =
+        StreamController<PlayerIdentifier<double>>.broadcast();
+    await AudioWaveformsInterface.instance.setMethodCallHandler();
     isInitialised = true;
   }
 
@@ -32,8 +36,17 @@ class PlatformStreams {
   Stream<PlayerIdentifier<PlayerState>> get onPlayerStateChanged =>
       _playerStateController.stream;
 
+  Stream<PlayerIdentifier<List<double>>> get onCurrentExtractedWaveformData =>
+      _extractedWaveformDataController.stream;
+
+  Stream<PlayerIdentifier<double>> get onExtractionProgress =>
+      _extractionProgressController.stream;
+
   late StreamController<PlayerIdentifier<int>> _currentDurationController;
   late StreamController<PlayerIdentifier<PlayerState>> _playerStateController;
+  late StreamController<PlayerIdentifier<List<double>>>
+      _extractedWaveformDataController;
+  late StreamController<PlayerIdentifier<double>> _extractionProgressController;
 
   void addCurrentDurationEvent(PlayerIdentifier<int> playerIdentifier) {
     if (!_currentDurationController.isClosed) {
@@ -47,9 +60,24 @@ class PlatformStreams {
     }
   }
 
-  void dispose() async {
-    await _currentDurationController.close();
-    await _playerStateController.close();
+  void addExtractedWaveformDataEvent(
+      PlayerIdentifier<List<double>> playerIdentifier) {
+    if (!_extractedWaveformDataController.isClosed) {
+      _extractedWaveformDataController.add(playerIdentifier);
+    }
+  }
+
+  void addExtractionProgress(PlayerIdentifier<double> progress)async{
+    if (!_extractionProgressController.isClosed) {
+      _extractionProgressController.add(progress);
+    }
+  }
+
+  void dispose() {
+    _currentDurationController.close();
+    _playerStateController.close();
+    _extractedWaveformDataController.close();
+    _currentDurationController.close();
     AudioWaveformsInterface.instance.removeMethodCallHandeler();
     isInitialised = false;
   }